refactor: colocate some types
