+++ apps/meteor/client/definitions/IAppExternalURL.ts
@@ -1,0 +1,3 @@
+export interface IAppExternalURL {
+	url: string;
+}
