+++ apps/meteor/client/definitions/ICategory.ts
@@ -1,0 +1,8 @@
+export interface ICategory {
+	createdDate: Date;
+	description: string;
+	hidden: boolean;
+	id: string;
+	modifiedDate: Date;
+	title: string;
+}
