+++ apps/meteor/client/views/marketplace/AppDetailsPage/AppDetailsPageTabs.tsx
@@ -2,7 +2,7 @@ import { Tabs, TabsItem } from '@rocket.chat/fuselage';
 import { usePermission, useRouter } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
-import type { ISettings } from '../../../apps/@types/IOrchestrator';
+import type { ISettings } from '../definitions/ISettings';
 
 export type AppDetailsPageTabsProps = {
 	context: string;
