+++ apps/meteor/client/views/marketplace/definitions/AppInfo.ts
@@ -1,7 +1,7 @@
 import type { IApiEndpointMetadata } from '@rocket.chat/apps-engine/definition/api';
 import type { AppScreenshot } from '@rocket.chat/core-typings';
 
-import type { ISettings } from '../../../apps/@types/IOrchestrator';
+import type { ISettings } from './ISettings';
 import type { App } from '../types';
 
 export type AppInfo = App & {
