+++ apps/meteor/client/views/marketplace/definitions/ISettings.ts
@@ -1,0 +1,5 @@
+import type { ISetting } from '@rocket.chat/apps-engine/definition/settings';
+
+export interface ISettings {
+	[key: string]: ISetting;
+}
